Refactor fragment names and create new packages.
