+++ app/src/main/java/com/rosseti/southevents/main/events/EventDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.rosseti.southevents
+package com.rosseti.southevents.main.events
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,19 +8,15 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.navigation.fragment.findNavController
-import kotlinx.android.synthetic.main.fragment_first.*
+import com.rosseti.southevents.R
 import kotlinx.android.synthetic.main.fragment_second.*
 
-/**
- * A simple [Fragment] subclass as the second destination in the navigation.
- */
-class SecondFragment : Fragment() {
+class EventDetailsFragment : Fragment() {
 
   override fun onCreateView(
     inflater: LayoutInflater, container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View? {
-    // Inflate the layout for this fragment
     return inflater.inflate(R.layout.fragment_second, container, false)
   }
 
@@ -35,9 +31,6 @@ class SecondFragment : Fragment() {
   }
 
   override fun onOptionsItemSelected(item: MenuItem): Boolean {
-    // Handle action bar item clicks here. The action bar will
-    // automatically handle clicks on the Home/Up button, so long
-    // as you specify a parent activity in AndroidManifest.xml.
     return when (item.itemId) {
       android.R.id.home -> {
         findNavController().navigate(R.id.action_SecondFragment_to_FirstFragment)
