+++ app/src/main/java/com/rosseti/southevents/main/events/EventLisFragment.kt
@@ -1,4 +1,4 @@
-package com.rosseti.southevents
+package com.rosseti.southevents.main.events
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,18 +8,15 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.core.os.bundleOf
 import androidx.navigation.fragment.findNavController
+import com.rosseti.southevents.R
 
-/**
- * A simple [Fragment] subclass as the default destination in the navigation.
- */
-class FirstFragment : Fragment() {
+class EventLisFragment : Fragment() {
 
   override fun onCreateView(
     inflater: LayoutInflater, container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View? {
-    // Inflate the layout for this fragment
-    return inflater.inflate(R.layout.fragment_first, container, false)
+    return inflater.inflate(R.layout.fragment_events, container, false)
   }
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
