+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,9 +7,9 @@
 
   <fragment
       android:id="@+id/FirstFragment"
-      android:name="com.rosseti.southevents.FirstFragment"
+      android:name="com.rosseti.southevents.main.events.EventLisFragment"
       android:label="@string/first_fragment_label"
-      tools:layout="@layout/fragment_first">
+      tools:layout="@layout/fragment_events">
 
     <action
         android:id="@+id/action_FirstFragment_to_SecondFragment"
@@ -17,7 +17,7 @@
   </fragment>
   <fragment
       android:id="@+id/SecondFragment"
-      android:name="com.rosseti.southevents.SecondFragment"
+      android:name="com.rosseti.southevents.main.events.EventDetailsFragment"
       android:label="@string/second_fragment_label"
       tools:layout="@layout/fragment_second">
